requirements.txt
langchain
langchain-community
langgraph
langchain-openai
chromadb
pypdf
streamlit
python-dotenv